Split renfxx.h
There were a number of minor issues with renfxx.h:

* IDEs and tools such as standardese get confused by the .h ending and
  assume that this is a C header. There's always ways to tell these tools
  but it's easier to just change the suffix.
* It is not good for build times to always include all of e-antic. We
  already had renfxx_fwd.h (now renamed to renfxx_fwd.hpp) for this but
  now one can also choose to only include the parent or the element of
  e-antic as needed.

I sneaked in an ODR violation fix in cppyy.h (now renamed to cppyy.hpp.)
A non-template definition in a header file should be marked inline so it
can be deduplicated.
