+++ doc/news/header-split.rst
@@ -1,0 +1,4 @@
+**Changed:**
+
+* Split `renfxx.h` into `renf_elem_class.hpp` and `renf_class.hpp` headers. The
+  original header still exists and simply includes the two novel headers.
